+++ lib/query/core/zen_cancel_token.dart
@@ -1,4 +1,4 @@
-import '../core/zen_logger.dart';
+import '../../core/zen_logger.dart';
 
 /// A token that can be used to signal cancellation to async operations.
 ///